+++ Tests/Public/Record/PrimaryKeyHiddenRowIDTests.swift
@@ -629,9 +629,9 @@ class PrimaryKeyHiddenRowIDTests : GRDBTestCase {
     func testExistsWithNilPrimaryKeyReturnsFalse() {
         assertNoError {
             let dbQueue = try makeDatabaseQueue()
-            dbQueue.inDatabase { db in
+            try dbQueue.inDatabase { db in
                 let record = Person(id: nil, name: "Arthur")
-                XCTAssertFalse(record.exists(db))
+                XCTAssertFalse(try record.exists(db))
             }
         }
     }
@@ -639,9 +639,9 @@ class PrimaryKeyHiddenRowIDTests : GRDBTestCase {
     func testExistsWithNotNilPrimaryKeyThatDoesNotMatchAnyRowReturnsFalse() {
         assertNoError {
             let dbQueue = try makeDatabaseQueue()
-            dbQueue.inDatabase { db in
+            try dbQueue.inDatabase { db in
                 let record = Person(id: 123456, name: "Arthur")
-                XCTAssertFalse(record.exists(db))
+                XCTAssertFalse(try record.exists(db))
             }
         }
     }
@@ -652,7 +652,7 @@ class PrimaryKeyHiddenRowIDTests : GRDBTestCase {
             try dbQueue.inDatabase { db in
                 let record = Person(name: "Arthur")
                 try record.insert(db)
-                XCTAssertTrue(record.exists(db))
+                XCTAssertTrue(try record.exists(db))
             }
         }
     }
@@ -664,7 +664,7 @@ class PrimaryKeyHiddenRowIDTests : GRDBTestCase {
                 let record = Person(name: "Arthur")
                 try record.insert(db)
                 try record.delete(db)
-                XCTAssertFalse(record.exists(db))
+                XCTAssertFalse(try record.exists(db))
             }
         }
     }